position int to decimal

+++ UserService/Controllers/ViewModels/Common/Position.cs
@@ -2,9 +2,9 @@ namespace UserService.Controllers.ViewModels.Common
 {
     public class Position
     {
-        public int X { set; get; }
-        public int Y { set; get; }
-        public int Z { set; get; }
+        public decimal X { set; get; }
+        public decimal Y { set; get; }
+        public decimal Z { set; get; }
 
         public Domains.ValueObject.Position ToModel()
         {
