+++ UserService/Domains/ValueObject/Position.cs
@@ -6,13 +6,13 @@ namespace UserService.Domains.ValueObject
     public class Position
     {
         [BsonElement("x")] 
-        public int X { private set; get; }
+        public decimal X { private set; get; }
         [BsonElement("y")] 
-        public int Y { private set; get; }
+        public decimal Y { private set; get; }
         [BsonElement("z")] 
-        public int Z { private set; get; }
+        public decimal Z { private set; get; }
 
-        public Position(int x, int y, int z)
+        public Position(decimal x, decimal y, decimal z)
         {
             X = x;
             Y = y;
